requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-telegram-bot==21.5
google-api-python-client==2.140.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
openai==1.51.0
python-dotenv==1.0.1
aiohttp==3.10.5
python-multipart==0.0.9
apscheduler==3.10.4
jinja2==3.1.4
aiofiles==24.1.0
openpyxl==3.1.5
python-docx==1.1.2
PyPDF2==3.0.1
Pillow==10.4.0
gspread==6.1.2
aiosmtplib==3.0.1
sqlalchemy==2.0.32
aiosqlite==0.20.0
